Initial Commit
This commit includes the initial code, and some changes and comments added to script.js file.
In the first version, I copied and pasted Pseudocode from the class session, and I created 4 functions that generate each a random thing: a lowercase letter, an uppercase letter, a special character, and a number. The first version of such formulas and connected variables were commented out.